requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.103.2
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.15.1
idna==3.11
keras==2.13.1
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
ml-dtypes==0.2.0
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.24
pydantic_core==2.41.5
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.5.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
typing-inspection==0.4.2
typing_extensions==4.5.0
urllib3==2.6.2
uvicorn==0.38.0
Werkzeug==3.1.4
wrapt==1.14.2
